# docs/swagger.yaml
basePath: /
definitions:
  models.Movie:
    properties:
      actors:
        items:
          $ref: '#/definitions/models.MovieActor'
        type: array
      description:
        type: string
      id:
        type: integer
      mark:
        maximum: 10
        minimum: 1
        type: integer
      poster_url:
        type: string
      review:
        type: string
      status:
        enum:
        - watched
        - planned
        type: string
      title:
        type: string
      type:
        enum:
        - film
        - serial
        type: string
      year:
        type: integer
    required:
    - status
    - title
    - type
    - year
    type: object
  models.MovieActor:
    properties:
      bio_url:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      role:
        type: string
    required:
    - first_name
    - last_name
    - role
    type: object
  models.User:
    properties:
      avatar_url:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
      registered_date:
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    type: object
  models.UserSignInRequest:
    properties:
      email:
        default: marisasha228@bk.ru
        type: string
      password:
        default: "123"
        type: string
    type: object
host: localhost:8000
info:
  contact:
    email: support@ttl-checker.com
    name: API Support
  description: API для кино заметок
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: TTL Kinolog API
  version: "1.0"
paths:
  /api/movies/:
    get:
      consumes:
      - application/json
      description: Посмотреть все фильмы пользователя
      operationId: get-movies
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Посмотреть все фильмы
      tags:
      - movies
  /api/movies/{id}:
    get:
      consumes:
      - application/json
      description: Посмотреть фильм пользователя
      operationId: get-movie
      parameters:
      - description: Id movie
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Посмотреть фильм
      tags:
      - movies
  /api/movies/add:
    post:
      consumes:
      - application/json
      description: Добавляет новое кино
      operationId: add-movie
      parameters:
      - description: Информация о кино
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Movie'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Добавить кино
      tags:
      - movies
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: Проверка прав пользователя
      operationId: sign-in
      parameters:
      - description: Данные пользователя
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.UserSignInRequest'
      produces:
      - application/json
      responses: {}
      summary: Аутентификация пользователя
      tags:
      - auth
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: Создание нового пользователя
      operationId: sign-up
      parameters:
      - description: Данные пользователя
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses: {}
      summary: Регистрация пользователя
      tags:
      - auth
securityDefinitions:
  ApiKeyAuth:
    description: 'Введите: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"